requirements.txt
ibm-cloud-sdk-core
ibm-platform-services
ibm-vpc
requests
pytz==2020.1
boto3
python-dotenv
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0